addded styled_container to components
I noticed that I was frequently creating containers nested in material, so I created styled_container which is basically just that to reduce code duplication. refactored almost every page to include this widget, except for the first 3 onboarding pages because when i implemented it they changed alignment.

todo: do the same thing for styled_button.dart