Redid theme
Reformated from BLoC theme (It was too cumbersome). So I wrapped the whole app in an inherited widget and am now using that for theming.

Also, the StyledContainer was preventing some widgets from updating when a theme change takes place. So, I fixed that. Also just in general it's not a good design practice to wrap a lot of stateful widgets in other stateful widgets like that.