Fixed orientation issues, locked orientation
Fixed the landscape orientation issues and then I decided to only allow portrait on phones. A big problem coming up is gonna be scaling to tablets. Not sure how that is gonna happen,whether I will need to split the repositories or what. Will decide in the future.

Stock screen is giving me a weird error. WIll look into tomorrow.